requirements.txt

torch>=2.0.0
numpy>=1.24.0
PyYAML>=6.0.0
matplotlib>=3.5.0
scipy>=1.9.0
